DAO: little modf contentValues(); Shoptivity: lat/long to Double; GetAllIntImpl: modified fun parseStringToDouble

+++ domain/src/main/java/com/rodrigolc/madridshops/domain/model/Shoptivity.kt
@@ -11,8 +11,8 @@ data class Shoptivity(
         val image: String,
         val logo: String,
         val openingHours: String,
-        val longitude: Float?,
-        val latitude: Float?,
+        val longitude: Double?,
+        val latitude: Double?,
         val type: String
         ) {
     init {
